Clear up file naming a big
