+++ External/FEXCore/Source/Utils/ELFContainer.cpp
