+++ Source/Tests/CMakeLists.txt
@@ -2,7 +2,7 @@ add_subdirectory(LinuxSyscalls)
 
 set(LIBS FEXCore Common CommonCore FEX_jemalloc)
 
-add_executable(FEXLoader ELFLoader.cpp)
+add_executable(FEXLoader FEXLoader.cpp)
 target_include_directories(FEXLoader PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/Source/)
 target_include_directories(FEXLoader PRIVATE ${CMAKE_BINARY_DIR}/generated)
